Fix burnup-dependent reaction rates (#911)
